First pass of complete social network
I've fixed the issue with the loop.  I'm not exactly sure why this is
but stata had issues with line 112.  When not defined as a "long" this
lead to issues where SN_hhid2 was misidentified.

Outside of that -- I changed the structure to load the data each time as
opposed to using Restore/Preserve commands.  It's a bit less efficient
but helps with debugging.